refactor: reworked observation
