+++ .package.resolved
@@ -41,7 +41,7 @@
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/Infomaniak/ios-core",
       "state" : {
-        "revision" : "cf001ba55dbf5f152353d66b55cd46350bd4b895"
+        "revision" : "e5fe8e03aa06375f20c8e40f4fae3a6af6e4d75e"
       }
     },
     {
